Add BD migrations via Flyway(doesnt work) Add profile page and ability to change password, change email

+++ src/main/java/com/letscode/mvcGuide/domain/Message.java
@@ -5,8 +5,8 @@ import javax.persistence.*;
 @Entity // This tells Hibernate to make a table out of this class
 public class Message {
     @Id
-    @GeneratedValue(strategy= GenerationType.AUTO)
-    private Integer id;
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
 
     private String text;
     private String tag;
@@ -29,11 +29,11 @@ public class Message {
     public Message() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
